+++ CentroLink/CentroLink/PromoTicketSetupModule/DatabaseEntities/PromoTicketSetupEventTypes.cs
@@ -7,7 +7,7 @@
         PromoTicketAddedFailed = 1119,
         PromoTicketModified = 1120,
         PromoTicketModifiedFailed = 1121,
-        PromoTicketDeleted=1122,
-        PromoTicketDeletedFailed=1123
+        PromoTicketDeleted = 1122,
+        PromoTicketDeletedFailed = 1123
     }
 }